refactor: move Context to context module
